+++ index.mjs
@@ -1,5 +1,5 @@
-import createTempDir from '../private/createTempDir.mjs';
-import fsPathRemove from '../private/fsPathRemove.mjs';
+import createTempDir from './private/createTempDir.mjs';
+import fsPathRemove from './private/fsPathRemove.mjs';
 
 /**
  * Asynchronously creates a disposable directory in the OS temporary directory
@@ -8,13 +8,9 @@ import fsPathRemove from '../private/fsPathRemove.mjs';
  * @name disposableDirectory
  * @param {Function} callback A sync or async callback that receives the temporary directory path.
  * @returns {Promise<void>} Resolves once the callback is done and the disposable directory is deleted.
- * @example <caption>Ways to `import`.</caption>
+ * @example <caption>How to `import`.</caption>
  * ```js
- * import { disposableDirectory } from 'disposable-directory';
- * ```
- *
- * ```js
- * import disposableDirectory from 'disposable-directory/public/disposableDirectory.mjs';
+ * import disposableDirectory from 'disposable-directory';
  * ```
  * @example <caption>How to use a disposable directory.</caption>
  * ```js