+++ public/index.mjs
@@ -1,1 +1,0 @@
-export { default as disposableDirectory } from './disposableDirectory.mjs';