+++ test/index.mjs
@@ -1,8 +1,8 @@
 import TestDirector from 'test-director';
+import testDisposableDirectory from './disposableDirectory.test.mjs';
 import testFsPathExists from './fsPathExists.test.mjs';
 import testCreateTempDir from './private/createTempDir.test.mjs';
 import testFsPathRemove from './private/fsPathRemove.test.mjs';
-import testDisposableDirectory from './public/disposableDirectory.test.mjs';
 
 const tests = new TestDirector();